Make the diff easier to read?
